improved OOP in code
