+++ src/main/java/duke/TaskList.java
@@ -1,8 +1,5 @@
 package duke;
 import java.io.Serializable;
-import java.time.DayOfWeek;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 
 /**
@@ -84,6 +81,30 @@ public class TaskList implements Serializable {
         return stringBuilder.toString();
     }
     /**
+     * For an input such as 'todo borrow book', letter is 'T' and string is 'borrow book'
+     *
+     * @param letter the letter corresponding to the first letter of the duke.Task
+     * @param string the string corresponding to the chunk of text after the word todo, deadline, or event
+     */
+    public String addTask(String letter, String string) throws DukeException {
+        if (letter.equals("T")) {
+            tasks.add(ToDo.newToDo(string));
+        }
+        if (letter.equals("D")) {
+            tasks.add(Deadline.newDeadline(getDescription(string), getBy(string)));
+        }
+        if (letter.equals("E")) {
+            tasks.add(Event.newEvent(getDescription(string), getFrom(string), getTo(string)));
+        }
+
+        int tasksSize = tasks.size();
+        StringBuilder stringBuilder = new StringBuilder();
+        stringBuilder.append(indent + "Got it. I've added this task:\n");
+        stringBuilder.append(megaIndent + tasks.get(tasksSize - 1).toString() + "\n");
+        stringBuilder.append(indent + "Now you have " + tasksSize + " tasks in the list.\n");
+        return stringBuilder.toString();
+    }
+    /**
      * This method encapsulates the functionality of marking a task as completed or not
      * For example, the input 'mark 1' will mark the duke.Task at position 0 at the TaskArray as 'marked'
      * @param string the input string
@@ -136,8 +157,7 @@ public class TaskList implements Serializable {
     public static String getBy(String string) throws DukeException {
         String slash = "/";
         int first = string.indexOf(slash);
-        int second = first + 3;
-        if (first == -1 || !string.substring(first, second).equals("/by")) {
+        if (first == -1 || !string.substring(first + 1, first + 3).contains("by")) {
             throw new DukeException("You need to add a by timing!");
         }
         return string.substring(first + 4); // returns "Sunday"
@@ -182,30 +202,6 @@ public class TaskList implements Serializable {
         return string.substring(secondSlash + 4);
     }
     /**
-     * For an input such as 'todo borrow book', letter is 'T' and string is 'borrow book'
-     *
-     * @param letter the letter corresponding to the first letter of the duke.Task
-     * @param string the string corresponding to the chunk of text after the word todo, deadline, or event
-     */
-    public String addTask(String letter, String string) throws DukeException {
-        if (letter.equals("T")) {
-            tasks.add(new ToDo(string));
-        }
-        if (letter.equals("D")) {
-            tasks.add(new Deadline(getDescription(string), convertToLocalDateTime(getBy(string))));
-        }
-        if (letter.equals("E")) {
-            tasks.add(new Event(getDescription(string), getFrom(string), getTo(string)));
-        }
-
-        int tasksSize = tasks.size();
-        StringBuilder stringBuilder = new StringBuilder();
-        stringBuilder.append(indent + "Got it. I've added this task:\n");
-        stringBuilder.append(megaIndent + tasks.get(tasksSize - 1).toString() + "\n");
-        stringBuilder.append(indent + "Now you have " + tasksSize + " tasks in the list.\n");
-        return stringBuilder.toString();
-    }
-    /**
      * This method encapsulates deleting of a task from TaskArray
      * For example, the input 'delete 3' will delete the duke.Task at position 2 of TaskArray
      *
@@ -225,128 +221,5 @@ public class TaskList implements Serializable {
             return stringBuilder.toString();
         }
     }
-    /**
-     * A function that takes in the by part of a duke.Deadline duke.Task, and converts it to a LocalDateTime
-     * For example, the input 'Sunday 1700' will return the corresponding LocalDateTime
-     *
-     * @param string the by part of the duke.Deadline duke.Task
-     * @return the LocalDateTime corresponding to the duke.Deadline
-     * @throws DukeException if a specific time in 24hr format is not put
-     */
-    public static LocalDateTime convertToLocalDateTime(String string) throws DukeException {
-        if (string.indexOf('/') != -1) {
-            if (string.lastIndexOf('/') + 5 == string.length()) { // "2/12/2019 1800"
-                throw new DukeException("put in a time pls");
-            }
-            LocalDateTime dateTime = parseDateTime(string, '/');
-            return dateTime;
-        } else if (string.indexOf('-') != -1) { //
-            if (string.lastIndexOf('-') + 3 == string.length()) { // "2019-10-15 1800"
-                throw new DukeException("put in a time pls");
-            }
-            LocalDateTime dateTime = parseDateTime(string, '-');
-            return dateTime;
-        } else { // "Mon 1800"
-            // problem 1: date does not overflow to next month
-            // problem 2: it goes backwards in day
-            String[] parts = string.split(" ");
-            if (parts.length == 1) {
-                throw new DukeException("put in a time pls");
-            }
-            String dayPart = parts[0];
-            String timePart = parts[1];
 
-            int year = LocalDate.now().getYear();
-            int month = LocalDate.now().getMonth().getValue();
-            int daysToAdd = -LocalDateTime.now().getDayOfWeek().compareTo(getDayOfWeek(dayPart.toUpperCase()));
-            int date = LocalDate.now().getDayOfMonth() + daysToAdd;
-
-            int hour = Integer.parseInt(timePart.substring(0, 2));
-            int minute = Integer.parseInt(timePart.substring(2, 4));
-
-            LocalDate temp = LocalDate.of(year, month, 1);
-            // temp LocalDate to obtain the maximum no. of days in that month
-            int maxDaysOfMonth = temp.lengthOfMonth();
-
-            if (date > maxDaysOfMonth) {
-                // Date overflows, adjust LocalDateTime to the next month
-                return LocalDateTime.of(year, month + 1, date - maxDaysOfMonth, hour, minute);
-            } else {
-                return LocalDateTime.of(year, month, date, hour, minute);
-            }
-        }
-    }
-    /**
-     * A function that helps convert a string to a LocalDateTime
-     *
-     * @param input the by part of the duke.Deadline duke.Task, e.g. "2/12/2019 1800"
-     * @param c whether the duke.Deadline is put in a '-' format or '/' format
-     * @return a LocalDateTime
-     * @throws DukeException if a specific time in 24hr format is not put
-     */
-    public static LocalDateTime parseDateTime(String input, char c) throws DukeException {
-        String[] parts = input.split(" ");
-        if (parts.length != 2) {
-            throw new DukeException("put in a time pls");
-        }
-
-        String datePart = parts[0];
-        String timePart = parts[1];
-
-        String[] dateComponents;
-        if (c == '/') {
-            dateComponents = datePart.split("/");
-        } else {
-            // c == '-'
-            dateComponents = datePart.split("-");
-        }
-
-        if (dateComponents.length != 3) {
-            throw new IllegalArgumentException("Invalid date format");
-        }
-
-        int date = Integer.parseInt(dateComponents[0]);
-        int month = Integer.parseInt(dateComponents[1]);
-        int year = Integer.parseInt(dateComponents[2]);
-        int hour = Integer.parseInt(timePart.substring(0, 2));
-        int minute = Integer.parseInt(timePart.substring(2, 4));
-
-        return LocalDateTime.of(year, month, date, hour, minute);
-    }
-
-    /**
-     * A function that takes in a user input that is the day of the week and returns the
-     * corresponding DayOfWeek
-     *
-     * @param string the user input that is a day of the week, e.g. "sun", "Tuesday", "Mon"
-     * @return the DayOfWeek as an enum
-     */
-    public static DayOfWeek getDayOfWeek(String string) {
-        DayOfWeek result;
-        String day = string.substring(0, 3);
-        switch(day) {
-        case "MON":
-            result = DayOfWeek.MONDAY;
-            break;
-        case "TUE":
-            result = DayOfWeek.TUESDAY;
-            break;
-        case "WED":
-            result = DayOfWeek.WEDNESDAY;
-            break;
-        case "THU":
-            result = DayOfWeek.THURSDAY;
-            break;
-        case "FRI":
-            result = DayOfWeek.FRIDAY;
-            break;
-        case "SAT":
-            result = DayOfWeek.SATURDAY;
-            break;
-        default:
-            // case "SUN"
-            result = DayOfWeek.SUNDAY;
-        }
-        return result;
-    }
 }
