+++ src/main/java/duke/ToDo.java
@@ -4,7 +4,7 @@ package duke;
  * one of the subclasses of a Task
  */
 public class ToDo extends Task {
-    public ToDo(String description) {
+    private ToDo(String description) {
         super(description);
     }
     @Override
@@ -15,4 +15,9 @@ public class ToDo extends Task {
     public String toFileString() {
         return "T" + super.toFileString();
     }
+
+    public static ToDo newToDo(String description) {
+        return new ToDo(description);
+    }
+
 }
