+++ src/main/resources/view/DialogBox.fxml
@@ -9,7 +9,7 @@
          prefWidth="400.0" type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/17"
          xmlns:fx="http://javafx.com/fxml/1">
     <children>
-        <Label fx:id="dialog" text="Label" wrapText="true" />
+        <Label fx:id="dialog" text="Label" wrapText="true" styleClass="label"/>
         <ImageView fx:id="displayPicture" fitHeight="99.0" fitWidth="99.0" pickOnBounds="true"
                    preserveRatio="true" />
     </children>
