+++ src/main/styles.css
@@ -1,0 +1,19 @@
+/* Style for the user's dialog box */
+.user {
+    -fx-background-color: lightblue; /* Set the background color */
+    -fx-border-color: blue; /* Set the border color */
+    -fx-border-width: 2px; /* Set the border width */
+}
+
+/* Style for Duke's dialog box */
+.duke {
+    -fx-background-color: lightgreen; /* Set the background color */
+    -fx-border-color: green; /* Set the border color */
+    -fx-border-width: 2px; /* Set the border width */
+}
+
+.label {
+    -fx-background-color: lightgreen; /* Set the background color */
+    -fx-border-color: green; /* Set the border color */
+    -fx-border-width: 2px; /* Set the border width */
+}
