+++ src/test/java/duke/ToDoTest.java
@@ -7,12 +7,12 @@ public class ToDoTest {
 
     @Test
     public void testToString() {
-        assertEquals("[T][ ] borrow book", new ToDo("borrow book").toString());
+        assertEquals("[T][ ] borrow book", ToDo.newToDo("borrow book").toString());
     }
 
     @Test
     public void testToFileString() {
-        assertEquals("T | O | borrow book", new ToDo("borrow book").toFileString());
+        assertEquals("T | O | borrow book", ToDo.newToDo("borrow book").toFileString());
     }
 
 }
